+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+## [Unreleased]
+
+- Add `COMMAND INFO` command (ACL categories, tips, key specifications and subcommands are not yet supported)
+
 ## [v0.1.0] - 2023-02-26
 
 - Add `PING` command